feat [pkg]: move historic package from internal
this will help us reconstruct the BOM from tiles on pivnet
